Projeto_ADC_esp32/flask_server/requirements.txt
flask==2.0.1
flask-socketio==5.1.1
pyserial==3.5
numpy==1.21.2
scipy==1.7.1